expose A11y metadata for use externally (e.g. Ace App)
